requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
requests
spotipy
SQLAlchemy
